Revert "move Workspace object to test file"
This reverts commit e3ae1dbc
